Ensure relative paths in /web for portability
Summary of changes:
- Created `web/src/utils/paths.ts` containing `getRelativePath` utility to convert root-relative paths to relative ones based on current URL depth.
- Refactored `web/src/hooks/useRelativePath.ts` to leverage the shared utility.
- Updated `web/src/api/client.ts` to use relative paths for redirects to the Legacy UI (e.g., `/google/login.php`) when a 401 error occurs, ensuring the app remains portable when served from subdirectories.
- Maintained absolute paths for API requests (e.g., `/api/`, `/user.php`, etc.) as explicitly requested to avoid breakages in proxy/routing logic.
- Verified that all internal Next-Gen UI links continue to use relative paths via the `rel()` helper.
- Validated the changes by running `npm run build` and `npx tsc` in the `web` directory.
